+++ CHANGELOG.md
@@ -31,6 +31,8 @@ All notable changes to this project will be documented in this file. The format 
   unintentionally.
 - Fixed a bug where line charts would sometimes be rendered without goal lines.
 - Added a missing Firestore index to make the admin feasibly fast again.
+- Fixed a bug where saving a placeholder OKR period without changes would result
+  in it being set to unix epoch.
 
 ### Security